requirements.txt
google-generativeai==0.8.5
unstructured==0.17.2
opensearch-py==3.0.0
tiktoken==0.9.0
gradio==5.34.2
langchain==0.3.26